requirements.txt
requests==2.28.2
beautifulsoup4==4.12.0
pandas==1.5.3
qdrant-client==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
urllib3==1.26.9
tqdm==4.65.0
openai==0.27.2
numpy==1.24.2